src/key_frame_selection_LMSKE/requirements_lmske.txt
numpy==1.26.4
scipy>=1.10,<1.12
tensorflow==2.5.0
torch>=1.12,<2.1
torchvision
torchaudio
opencv-python-headless
ffmpeg-python
pillow
open_clip_torch
scikit-learn==1.2.2
